requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
distlib==0.3.7
Django==4.0.10
django-celery-beat==2.3.0
django-environ==0.10.0
django-extensions==3.2.3
django-timezone-field==6.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
factory-boy==3.2.1
Faker==19.3.0
filelock==3.12.2
hiredis==2.0.0
identify==2.5.27
iniconfig==2.0.0
kombu==5.3.1
mysql-connector-python==8.1.0
mysqlclient==2.2.0
nodeenv==1.8.0
packaging==23.1
platformdirs==3.10.0
pluggy==1.2.0
pre-commit==3.3.3
prompt-toolkit==3.0.39
protobuf==4.21.12
PyMySQL==1.1.0
pytest==7.4.0
pytest-sugar==0.9.5
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==4.3.4
redis-server==6.0.9
six==1.16.0
sqlparse==0.4.4
termcolor==2.3.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
virtualenv==20.24.3
wcwidth==0.2.6
wrapt==1.15.0
